refactor(core): pull duplicated iterator boilerplate up into FileFormatUtils
Three methods on the FileFormatUtils contract had byte-identical bodies in
ParquetUtils, OrcUtils, and LanceUtils; convert them from abstract to
concrete on the base class:

- getHoodieKeyIterator(storage, filePath) delegates to the 4-arg overload
- fetchRecordKeysWithPositions(storage, filePath) delegates to the 4-arg one
- fetchRecordKeysWithPositions(storage, filePath, keyGen, partitionPath)
  wraps getHoodieKeyIterator in a position-tracking CloseableMappingIterator

OrcUtils keeps its 4-arg fetch override (adds an existence guard) and
HFileUtils keeps its UnsupportedOperationException overrides so behavior is
unchanged everywhere; only the redundant Parquet/Orc/Lance copies are gone.
